chỉnh sửa bill và chỉnh sửa số lượng trong giỏ hàng

+++ src/main/java/com/example/sd40/entity/Gio_hang/GioHangDetail.java
@@ -28,7 +28,7 @@ public class GioHangDetail {
 
     private int quantity;
 
-    private int price;
+    private BigDecimal price;
 
     @ManyToOne
     @JoinColumn(name = "id_gio_hang",referencedColumnName = "id")
